DB integration

+++ codework-api/src/main/java/com/codework/repository/ChallengeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.codework.entity.Challenge;
 
-public interface ChallengeRepository extends MongoRepository<Challenge, Integer>{
+public interface ChallengeRepository extends MongoRepository<Challenge, Long>{
 
 }
